Create jessica_biel_naked_in_my_bed.c
The title of the exploit wasn't by me, and it wasn't written by me either. This piece of (modified) code is intended to have it compile and run smoothly on Kali Linux. This is so that others do not spend ages tearing their hair out, setting fire to other peoples houses, and wrecking a Nokia phone to get this exploit working. 
You should also get and compile pagesize.c on the target machine, take the output, and change the value of PAGE_SIZE. 
Don't hesitate to contact me!

s1xpence